+++ src/CustomSettingLayer.cpp
@@ -7,63 +7,15 @@ using namespace geode::prelude;
 
 $execute {
     new EventListener<EventFilter<ModPopupUIEvent>>(+[](ModPopupUIEvent* event) {
-        if (event->getModID() == "omgrod.geodify") {
+        if (event->getModID() == Mod::get()->getID()) {
             auto popup = event->getPopup();
             
             if (popup) {
                 if (CCMenuItemSpriteExtra* btn = typeinfo_cast<CCMenuItemSpriteExtra*>(popup->getChildByIDRecursive("settings-button"))) {
-                    btn->m_pfnSelector = menu_selector(GYSettingSelectLayer::settingsBtn); // this code sucks - it keeps crashing. im sorry if creating a fake button is a bit too much
-                    
-                    // CCSprite* settings = CCSprite::create("geode.loader/settings.png");
-                    // CircleButtonSprite* settingsBtn = nullptr;
-
-                    // if (Loader::get()->getLoadedMod("geode.loader")->getSettingValue<bool>("enable-geode-theme")) {
-                    //     settingsBtn = CircleButtonSprite::create(settings, CircleBaseColor::DarkPurple, CircleBaseSize::SmallAlt);
-                    // } else {
-                    //     settingsBtn = CircleButtonSprite::create(settings, CircleBaseColor::Green, CircleBaseSize::SmallAlt);
-                    // }
-
-                    // btn->setVisible(false);
-
-                    // auto newBtn = CCMenuItemSpriteExtra::create(settingsBtn, nullptr, );
-                    // newBtn->setID("settings-button-geodify");
-                    // newBtn->setPosition(btn->getPosition());
-                    // newBtn->setAnchorPoint(btn->getAnchorPoint());
-                    // newBtn->setScale(btn->getScale());
-                    // btn->getParent()->addChild(newBtn);
+                    btn->m_pfnSelector = menu_selector(GYSettingSelectLayer::settingsBtn);
                 }
             }
         }
         return ListenerResult::Propagate;
     });
 }
-
-/// @brief The following code is for testing purposes only, and adds a Geodify settings button to the main menu.
-
-/*
-#include <Geode/modify/MenuLayer.hpp>
-
-class $modify(MyMenuLayer, MenuLayer) {
-public:
-    bool init() {
-        if (!MenuLayer::init())
-            return false;
-
-        auto settingsBtn = CCMenuItemSpriteExtra::create(
-            CircleButtonSprite::create(
-                CCSprite::create("geode.loader/settings.png"),
-                CircleBaseColor::DarkPurple,
-                CircleBaseSize::Medium
-            ),
-            this,
-            menu_selector(GYSettingSelectLayer::settingsBtn)
-        );
-
-        settingsBtn->setID("settings-button-geodify");
-
-        auto menu = this->getChildByID("bottom-menu");
-        menu->addChild(settingsBtn);
-
-        return true;
-    }
-}*/